==> Programming/Digital House (Python)/Project 1 - Transport and real estate/2) Private transport.ipynb <==
# 1) Preparación previa

### Carga de librerías

# Se debe instalar mpu
import mpu #pip install git+https://github.com/MartinThoma/mpu.git
import pandas as pd
import re
import numpy as np
import seaborn as sns

### Funciones a usar luego

# La siguiente función nos permite definir la distancia en latitud y longitud respecto de los puentes:
def distancia(lat, long):
    aux = []
    for index, row in puentes_pilar.iterrows():
        aux.append(
            mpu.haversine_distance(
                (row['Latitud'], row['Longitud']), 
                (lat, long)
            )
        )
    return min(aux)

# La siguiente función nos permite aplicar íconos a los marcadores según su quantil:
def marca_propiedad(lat, long, quantile):
    if quantile == 0:
        icono_prop = icon0
    elif quantile == 1:
        icono_prop = icon1
    elif quantile == 2:
        icono_prop = icon2  
    else: 
        icono_prop = icon3
    marker = Marker(location=(lat,long), draggable=False, icon=icono_prop)
    basic_map.add_layer(marker)

### Lectura del dataset limpiado en la notebook de análisis

data = pd.read_csv('DF_Final.csv')
data.head()

### Filtro del DF

# Máscara para obtener los pertenecientes a Pilar
mascara_pilar = data.Partido == 'Pilar'
df_solo_pilar = data.loc[mascara_pilar, :]
df_solo_pilar.head()

# Máscara para obtener las viviendas que no son countries
pilar_partido = df_solo_pilar.isin({'Barrios':['Pilar', 'Del Viso', 'Derqui', 'Villa Rosa', 'Villa Stolfi', 'Manuel Alberti', 'Manzanares'] })
pilar_sincountry = df_solo_pilar[pilar_partido['Barrios']]
pilar_sincountry.head()

# Máscara para obtener las viviendas que son countries
pilar_country = df_solo_pilar[~pilar_partido ['Barrios']]
pilar_country.head()

# Máscara para eliminar valores menores a 1
precio_pilar_sin_country = pilar_sincountry['price_aprox_usd'] > 0 # or np.isNan
df_ppsc = pilar_sincountry.loc[precio_pilar_sin_country, :]
# Verificación de valores NaN -> ppsc = partido pilar sin country
df_ppsc.sort_values(by = 'price_aprox_usd', ascending = False)

### Eliminación de los valores nulos

df_ppsc_latlon = df_ppsc.dropna(subset=['lat', 'lon'])
df_ppsc_latlon.shape

# Se procede también con la columna 'Precio_USD_por_M2'
df_ppsc_latlon_con_precio = df_ppsc_latlon[df_ppsc_latlon['Precio_USD_por_M2'].notna()]
df_ppsc_latlon_con_precio

# 2) Análisis del DF sin countries

### Definición de cuantiles

df_ppsc_latlon_con_precio['quantile'] = pd.qcut(df_ppsc_latlon_con_precio['Precio_USD_por_M2'], 4, labels=False)
df_ppsc_latlon_con_precio.round(3)

### Métricas generales del DF sin countries

medidas_ppsc = df_ppsc_latlon_con_precio[["Precio_USD_por_M2"]].describe().round(2)
medidas_ppsc

### Agrupación según barrios

df_ppsc_agrup = df_ppsc_latlon_con_precio.groupby('Barrios')
df_ppsc_agrup

### Métricas del DF agrupado sin countries

medidas_ppsc_agrup = df_ppsc_agrup[["Precio_USD_por_M2"]].describe().round(2)
medidas_ppsc_agrup

# Orden según la media
medidas_ppsc_agrup.sort_values(by=[('Precio_USD_por_M2', 'mean')], ascending = False)

# 3) Preparación del DF con countries

### Eliminación de valores inconsistentes

# Máscara para eliminar valores menores a 1 | ppc = partido pilar countries
precio_pilar_country = pilar_country['price_aprox_usd'] > 1
df_ppc = pilar_country.loc[precio_pilar_country, :]
# Verificación de valores NaN
df_ppc.sort_values(by = 'price_aprox_usd', ascending = False)

# Eliminación de los nulos de la columna LAT y LON
df_ppc_latlon = df_ppc.dropna(subset=['lat', 'lon'])
df_ppc_latlon.shape

### Creación de nuevo DF con puentes

# Ubicación de los puentes de la panamericana - ramal pilar
puentes_pilar = pd.DataFrame({
    'Nombre': ['Separacion', 'Tortugas', '26', 'Camaño', 'Palmas', '25', 'Manzanares','6'],
    'Latitud': [-34.468040, -34.448354, -34.440682, -34.436667, -34.443766, -34.449550, -34.441083, -34.395740],
    'Longitud': [-58.685014, -58.748577, -58.782740, -58.828145, -58.872848, -58.905720, -58.966898, -59.022886]
})

print(puentes_pilar)

### Cálculo de la distancia hacia los puentes

%%capture output
# Cálculo d la distancia de las propiedades hacia los puentes en una nueva columna
df_ppc_latlon['distancia'] = df_ppc_latlon.apply(lambda x: distancia(x['lat'], x['lon']), axis=1)
df_ppc_latlon

### Eliminación de los nulos del precio por metro cuadrado

df_ppc_latlon_con_precio = df_ppc_latlon[df_ppc_latlon['Precio_USD_por_M2'].notna()]
df_ppc_latlon_con_precio

### Eliminación de outliers

mask_latlon_con_precio = df_ppc_latlon_con_precio['distancia'] < 90
df_ppc_latlon_con_precio = df_ppc_latlon_con_precio.loc[mask_latlon_con_precio, :]

### Eliminación de barrios nulos

mask_nulos = df_ppc_latlon_con_precio['Barrios'] != ""
df_final_countries = df_ppc_latlon_con_precio.loc[mask_nulos, :]
df_final_countries

# 4) Análisis del DF con countries

### Armado de cuantiles

df_final_countries['quantile'] = pd.qcut(df_final_countries['Precio_USD_por_M2'], 4, labels=False)
df_final_countries.round(3)

### Métricas generales del DF con countries

medidas_ppc = df_final_countries[["Precio_USD_por_M2"]].describe().round(2)
medidas_ppc

### Agrupación según barrios

df_ppc_agrup = df_final_countries.groupby('Barrios')
df_ppc_agrup

### Métricas específicas del DF con countries agrupado

medidas_ppc_agrup = df_ppc_agrup[["Precio_USD_por_M2"]].describe().round(2)

# Orden según la media
medidas_ppc_agrup.sort_values(by=[('Precio_USD_por_M2', 'mean')], ascending = False)

# 4) Gráficos

### Preparación de las variables a usar

distancia_graf = df_ppc_latlon_con_precio['distancia']*1000
distancia_graf = distancia_graf.astype(int)    

Precio_USD_por_M2 = np.around(df_ppc_latlon_con_precio['Precio_USD_por_M2']) 
price_aprox_usd= np.around(df_ppc_latlon_con_precio['price_aprox_usd'])

b2=Precio_USD_por_M2.mean()
c2=distancia_graf.mean()
a2=price_aprox_usd.mean()

b3=Precio_USD_por_M2.median()
c3=distancia_graf.median()
a3=price_aprox_usd.median()

a3,b3,c3

### Gráfico bivariable de distancia vs precio por m² y precio total

with sns.axes_style('dark'):
    g =sns.jointplot( x=distancia_graf, y=Precio_USD_por_M2 , height=10, ratio=3 )
    f= sns.jointplot( x=distancia_graf, y=price_aprox_usd , height=10, ratio=3 )

### Gráfico de quantiles

quantile1 = np.around(df_final_countries['quantile'])
bx = sns.boxplot(x = quantile1, y = df_final_countries['distancia'])

ax = sns.boxplot(x = quantile1, y = Precio_USD_por_M2)

### Gráfico de cercanía a la panamericana

import folium

# Se crea un mapa
map = folium.Map(location=[-34.44349933885334, -58.86776001513485], zoom_start=13)

# Se agregan los íconos para los puentes
url = 'https://raw.githubusercontent.com/Agustin-Bulzomi/Projects/main/Programming/Digital%20House%20(Python)/Support%20Files/Project%201/%C3%8Dcono%20puente.png'
for index, row in puentes_pilar.iterrows():
    folium.Marker([row['Latitud'], row['Longitud']], icon=folium.features.CustomIcon(url, icon_size=(28, 28)), zindex_offset=2000).add_to(map)

# Se agregan los íconos para las casas
for index, row in df_final_countries.iterrows():
    if row['quantile'] == 0:
        folium.Marker([row['lat'], row['lon']], icon=folium.Icon(icon='home', color='green'), z_index_offset = -1000).add_to(map)
    elif row['quantile'] == 1:
        folium.Marker([row['lat'], row['lon']], icon=folium.Icon(icon='home', color='beige'), z_index_offset = -1000).add_to(map)
    elif row['quantile'] == 2:
        folium.Marker([row['lat'], row['lon']], icon=folium.Icon(icon='home', color='orange'), z_index_offset = -1000).add_to(map)
    else:
        folium.Marker([row['lat'], row['lon']], icon=folium.Icon(icon='home', color='red'), z_index_offset = -1000).add_to(map)

display(map)
