==> Programming/Digital House (Python)/Project 1 - Transport and real estate/3) Public transport.ipynb <==
# 1) Preparación previa

### Carga de librerías

import requests
from pandas import json_normalize
import mpu
import pandas as pd
import numpy as np
import folium

### Estaciones a usar luego

# Dirección web para acceder a los datos del gobierno de la ciudad:
url = 'https://cdn.buenosaires.gob.ar/datosabiertos/datasets/subte-estaciones/subte_estaciones.geojson'
response = requests.get(url)
dictr = response.json()

# Hacemos referencia a las estaciones dentro del conjunto de datos:
estaciones = dictr['features']
df = json_normalize(estaciones)

# Creamos la columna Longitud:
df['Longitud'] = df['geometry.coordinates'].apply(lambda x:x[0])

# Creamos la columna Latitud:
df['Latitud'] = df['geometry.coordinates'].apply(lambda x:x[1])

# Creamos la máscara para filtrar los de Linea A:
mascara_lineas_elegidas = df['properties.LINEA'].isin(['A', 'B', 'C', 'E', 'H']) 
lineas_elegidas = df.loc[mascara_lineas_elegidas, :]
lineas_elegidas = lineas_elegidas.drop(['type', 'properties.ID', 'properties.LINEA', 'geometry.type', 'geometry.coordinates'], axis = 1)
lineas_elegidas

### Funciones a usar luego

# La siguiente función nos permite definir la distancia en latitud y longitud respecto de las líneas de subte:
def distancia(lat, long):
    aux = []
    for index, row in lineas_elegidas.iterrows():
        aux.append(
            mpu.haversine_distance(
                (row['Latitud'], row['Longitud']), 
                (lat, long)
            )
        )
    return min(aux)

# La siguiente función nos permite limpiar la superficie según las inconsistencias entre la total y la cubierta.
# Se eligió dicha columna con respecto a 'surface_total_in_m2' ya que tenía menor cantidad de registros nulos
def limpieza_superficie(sup_total, sup_cubierta):
    if sup_total is not None and sup_cubierta is not None:
            if sup_total >= sup_cubierta:
                superficie = sup_total
            else: 
                superficie = np.NaN
    elif sup_total is not None:
            superficie = sup_total
    elif sup_cubierta is not None:
            superficie = sup_cubierta
    else: superficie = np.NaN             
                    
    return superficie

### Lectura del dataset original de Properati

df_properati = pd.read_csv("https://media.githubusercontent.com/media/Agustin-Bulzomi/Projects/main/Programming/Digital%20House%20(Python)/Support%20Files/Project%201/Properati.csv")

# 2) Limpieza del dataset

### Generación de DF solo para Capital

df_capital = df_properati[df_properati["state_name"] == "Capital Federal"]
df_capital.shape

### Corrección de ortografía de barrios y mapeo de barrios inexistentes

%%capture output
df_capital["place_name"].replace(['San Cristobal','Barrio Norte', 'Centro / Microcentro'], ['San Cristóbal','Recoleta', 'Monserrat'], inplace=True)

### Lectuura el archivo de Mapping de Barrios por Comuna

df_barrios_comuna = pd.read_csv("https://raw.githubusercontent.com/Agustin-Bulzomi/Projects/main/Programming/Digital%20House%20(Python)/Support%20Files/Project%201/Barrios_por_comuna.csv")

### Merge del DF de Capital con los barrios según comunas

df_capital_comuna = pd.merge(df_capital, df_barrios_comuna, how='left', left_on='place_name', right_on='Barrio')
del df_capital_comuna["description"]
del df_capital_comuna["title"]
del df_capital_comuna["properati_url"]
del df_capital_comuna["image_thumbnail"]

df_capital_comuna["Comuna"].isna().sum()

### Enumeración de los barrios por los que atraviesa la Linea A, elegida como la principal para analizar.

barrios_linea_a = pd.Series(['Monserrat', 'Balvanera', 'Almagro', 'Caballito'])
barrios_linea_a

### Filtro del DF de Capital según los barrios definidos

df_capital_barrios_subte = df_capital_comuna[df_capital_comuna['Barrio'].isin(barrios_linea_a)]
df_capital_barrios_subte.shape

### Eliminación de los registros sin coordenadas

df_latlon_notna = df_capital_barrios_subte[df_capital_barrios_subte['lat-lon'].notna()]

df_latlon_notna.shape

### Eliminación de los registros sin precios

df_price_notna = df_latlon_notna[df_latlon_notna['price_usd_per_m2'].notna()]
df_price_notna

### Limpieza de la superficie

# Se debe corregir inconsistencias entre la superficie total y la cubierta

%%capture output
df_price_notna["superficie"] = df_price_notna[['surface_total_in_m2','surface_covered_in_m2']].apply(lambda df_price_notna: limpieza_superficie(df_price_notna['surface_total_in_m2'],df_price_notna['surface_covered_in_m2']),axis=1)
df_price_notna.head(4)

# 3) Análisis de hipótesis

### Cálculo de las distancias entre cada propiedad y las 17 estaciones

%%capture output
df_price_notna['distancia'] = df_price_notna.apply(lambda x: distancia(x['lat'], x['lon']), axis=1)
df_price_notna

### Sampleo aleatorio para obtener una visualización más prolija

df_randomizado = df_price_notna.sample(n=100, random_state = 1)

### Definición de quantiles

df_randomizado['quantile'] = pd.qcut(df_randomizado['price_usd_per_m2'], 4, labels=False)
df_randomizado

### Visualización en un mapa

url = 'https://upload.wikimedia.org/wikipedia/commons/thumb/b/bf/Subte-logo.svg/1200px-Subte-logo.svg.png'

map = folium.Map(location=[-34.613, -58.41], zoom_start=14)

for index, row in lineas_elegidas.iterrows():
    marker_subte = folium.Marker([row['Latitud'], row['Longitud']], icon=folium.features.CustomIcon(url, icon_size=(28, 28))).add_to(map)

# Se agregan los íconos para las casas
for index, row in df_randomizado.iterrows():
    if row['quantile'] == 0:
        folium.Marker([row['lat'], row['lon']], icon=folium.Icon(icon='home', color='green')).add_to(map)
    elif row['quantile'] == 1:
        folium.Marker([row['lat'], row['lon']], icon=folium.Icon(icon='home', color='beige')).add_to(map)
    elif row['quantile'] == 2:
        folium.Marker([row['lat'], row['lon']], icon=folium.Icon(icon='home', color='orange')).add_to(map)
    else:
        folium.Marker([row['lat'], row['lon']], icon=folium.Icon(icon='home', color='red')).add_to(map)

display(map)


### Exportación del .CSV

df_randomizado.to_csv("df_randomizado_CABA.csv", index=False)
